improved UX in delete action

+++ src/pages/Profile.tsx
@@ -4,7 +4,8 @@ import { UpdateUserProfile } from '../components/Profile/UpdateUserProfile';
 import { useDispatch } from "react-redux";
 import { deleteUser } from '../redux/apiCall';
 import { RootStateOrAny, useSelector } from "react-redux";
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const Profile = () => {
 
@@ -13,7 +14,20 @@ const Profile = () => {
     const dispatch = useDispatch();
 
     const handleDelete = async () => {
-        dispatch(deleteUser(dispatch, _id));
+
+        Swal.fire({
+            title: 'Estas seguro que quieres eliminar tu cuenta?',
+            text: "Esta acción no es reversible",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, eliminar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteUser(dispatch, _id));
+            }
+        })
     }
 
     return (
@@ -25,7 +39,7 @@ const Profile = () => {
                             <UserProfile />
                             <div className="mt-6">
                                 <div className="profileContainer p-2">
-                                    <button className="p-2 m-2 bg-blue-200 hover:bg-blue-200/90 text-black cursor-pointer rounded-md" onClick={() => setIsEditing(!isEditing)}>Actualizar</button>
+                                    <button disabled={true} className="p-2 m-2 bg-blue-200 hover:bg-blue-200/90 text-black cursor-pointer rounded-md disabled:bg-gray-400 disabled:text-gray-300 disabled:cursor-not-allowed" onClick={() => setIsEditing(!isEditing)}>Actualizar</button>
                                     <Link to={"/"} className="p-2 m-2 bg-red-200 hover:bg-red-200/90 text-black cursor-pointer rounded-md" onClick={handleDelete}>Eliminar</Link>
                                 </div>
                             </div>
